docs/12-capstones/other/loop-analytics/requirements.txt
agate==1.6.1
appnope==0.1.2
asn1crypto==1.4.0
attrs==21.2.0
azure-common==1.1.27
azure-core==1.14.0
azure-storage-blob==12.8.1
Babel==2.9.1
backcall==0.2.0
backports-datetime-fromisoformat==1.0.0
boto3==1.20.42
botocore==1.23.42
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.7
dataclasses==0.6
dbt==0.21.1
dbt-bigquery==0.21.1
dbt-core==0.21.1
dbt-extractor==0.4.0
dbt-postgres==0.21.1
dbt-redshift==0.21.1
dbt-snowflake==0.21.1
decorator==4.4.2
google-api-core==1.23.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.3.1
google-cloud-core==1.4.4
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.52.0
greenlet==1.1.0
gspread==3.7.0
hologram==0.0.14
httplib2==0.20.2
idna==2.9
importlib-metadata==4.0.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
json-rpc==1.13.0
jsonschema==3.1.1
keyring==21.8.0
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.5
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
msrest==0.6.21
networkx==2.6.3
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.1
oscrypto==1.2.1
packaging==20.9
pandas==1.1.5
parsedatetime==2.6
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
proto-plus==1.19.8
protobuf==3.19.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.13.0
Pygments==2.8.1
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.8
s3transfer==0.5.0
six==1.16.0
snowflake-connector-python==2.5.1
SQLAlchemy==1.4.17
sqlalchemy-redshift==0.8.2
sqlparse==0.4.2
stripe==2.58.0
text-unidecode==1.3
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.7.0
